requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.3
geopandas==0.14.3
shapely==2.0.3
fiona==1.9.5
rasterio==1.3.9
pyproj==3.6.1
scipy==1.12.0
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
reportlab==4.1.0
httpx==0.26.0
aiofiles==23.2.1
requests==2.31.0
matplotlib==3.8.2
fpdf2==2.7.0
pillow==10.2.0
